Create first version of the library
