+++ examples/gradcam.py
@@ -1,0 +1,83 @@
+import argparse
+
+import cv2
+import numpy as np
+import torch
+from torchvision import models
+
+from pytorch_grad_cam.gradcam import GradCam, preprocess_image, GuidedBackpropReLUModel
+
+
+def get_args():
+    parser = argparse.ArgumentParser()
+    parser.add_argument('--use-cuda', action='store_true', default=False,
+                        help='Use NVIDIA GPU acceleration')
+    parser.add_argument('--image-path', type=str, default='./examples/both.png',
+                        help='Input image path')
+    args = parser.parse_args()
+    args.use_cuda = args.use_cuda and torch.cuda.is_available()
+    if args.use_cuda:
+        print("Using GPU for acceleration")
+    else:
+        print("Using CPU for computation")
+
+    return args
+
+
+def deprocess_image(img):
+    """ see https://github.com/jacobgil/keras-grad-cam/blob/master/grad-cam.py#L65 """
+    img = img - np.mean(img)
+    img = img / (np.std(img) + 1e-5)
+    img = img * 0.1
+    img = img + 0.5
+    img = np.clip(img, 0, 1)
+    return np.uint8(img * 255)
+
+
+def show_cam_on_image(img, mask):
+    heatmap = cv2.applyColorMap(np.uint8(255 * mask), cv2.COLORMAP_JET)
+    heatmap = np.float32(heatmap) / 255
+    cam = heatmap + np.float32(img)
+    cam = cam / np.max(cam)
+    return np.uint8(255 * cam)
+
+
+if __name__ == '__main__':
+    """ python pytorch_grad_cam.py <path_to_image>
+    1. Loads an image with opencv.
+    2. Preprocesses it for VGG19 and converts to a pytorch variable.
+    3. Makes a forward pass to find the category index with the highest score,
+    and computes intermediate activations.
+    Makes the visualization. """
+
+    args = get_args()
+
+    model = models.resnet50(pretrained=False)
+    grad_cam = GradCam(model=model, feature_module=model.layer4,
+                       target_layer_names=["2"], use_cuda=args.use_cuda)
+
+    img = cv2.imread(args.image_path, 1)
+    img = np.float32(img) / 255
+    # Opencv loads as BGR:
+    img = img[:, :, ::-1]
+    input_img = preprocess_image(img)
+
+    # If None, returns the map for the highest scoring category.
+    # Otherwise, targets the requested category.
+    target_category = None
+    grayscale_cam = grad_cam(input_img, target_category)
+
+    grayscale_cam = cv2.resize(grayscale_cam, (img.shape[1], img.shape[0]))
+    cam = show_cam_on_image(img, grayscale_cam)
+
+    gb_model = GuidedBackpropReLUModel(model=model, use_cuda=args.use_cuda)
+    gb = gb_model(input_img, target_category=target_category)
+    gb = gb.transpose((1, 2, 0))
+
+    cam_mask = cv2.merge([grayscale_cam, grayscale_cam, grayscale_cam])
+    cam_gb = deprocess_image(cam_mask * gb)
+    gb = deprocess_image(gb)
+
+    cv2.imwrite("cam.jpg", cam)
+    cv2.imwrite('gb.jpg', gb)
+    cv2.imwrite('cam_gb.jpg', cam_gb)
