+++ pytorch_grad_cam/__init__.py
@@ -1,0 +1,1 @@
+from pytorch_grad_cam.gradcam import GradCam, GuidedBackpropReLUModel
