+++ pytorch_grad_cam/gradcam.py
@@ -1,12 +1,11 @@
-import torch
-import argparse
 import cv2
 import numpy as np
 import torch
 from torch.autograd import Function
-from torchvision import models, transforms
+from torchvision import transforms
 
-class FeatureExtractor():
+
+class FeatureExtractor:
     """ Class for extracting activations and
     registering gradients from targetted intermediate layers """
 
@@ -28,7 +27,8 @@ class FeatureExtractor():
                 outputs += [x]
         return outputs, x
 
-class ModelOutputs():
+
+class ModelOutputs:
     """ Class for making a forward pass, and getting:
     1. The network output.
     2. Activations from intermeddiate targetted layers.
@@ -49,27 +49,22 @@ class ModelOutputs():
                 target_activations, x = self.feature_extractor(x)
             elif "avgpool" in name.lower():
                 x = module(x)
-                x = x.view(x.size(0),-1)
+                x = x.view(x.size(0), -1)
             else:
                 x = module(x)
 
         return target_activations, x
 
+
 def preprocess_image(img):
     normalize = transforms.Normalize(mean=[0.485, 0.456, 0.406],
-                                 std=[0.229, 0.224, 0.225])
+                                     std=[0.229, 0.224, 0.225])
     preprocessing = transforms.Compose([
         transforms.ToTensor(),
         normalize,
     ])
     return preprocessing(img.copy()).unsqueeze(0)
 
-def show_cam_on_image(img, mask):
-    heatmap = cv2.applyColorMap(np.uint8(255 * mask), cv2.COLORMAP_JET)
-    heatmap = np.float32(heatmap) / 255
-    cam = heatmap + np.float32(img)
-    cam = cam / np.max(cam)
-    return np.uint8(255 * cam)
 
 class GradCam:
     def __init__(self, model, feature_module, target_layer_names, use_cuda):
@@ -91,7 +86,7 @@ class GradCam:
 
         features, output = self.extractor(input_img)
 
-        if target_category == None:
+        if target_category is None:
             target_category = np.argmax(output.cpu().data.numpy())
 
         one_hot = np.zeros((1, output.size()[-1]), dtype=np.float32)
@@ -99,7 +94,7 @@ class GradCam:
         one_hot = torch.from_numpy(one_hot).requires_grad_(True)
         if self.cuda:
             one_hot = one_hot.cuda()
-        
+
         one_hot = torch.sum(one_hot * output)
 
         self.feature_module.zero_grad()
@@ -173,7 +168,7 @@ class GuidedBackpropReLUModel:
 
         output = self.forward(input_img)
 
-        if target_category == None:
+        if target_category is None:
             target_category = np.argmax(output.cpu().data.numpy())
 
         one_hot = np.zeros((1, output.size()[-1]), dtype=np.float32)
@@ -189,67 +184,3 @@ class GuidedBackpropReLUModel:
         output = output[0, :, :, :]
 
         return output
-
-def get_args():
-    parser = argparse.ArgumentParser()
-    parser.add_argument('--use-cuda', action='store_true', default=False,
-                        help='Use NVIDIA GPU acceleration')
-    parser.add_argument('--image-path', type=str, default='./examples/both.png',
-                        help='Input image path')
-    args = parser.parse_args()
-    args.use_cuda = args.use_cuda and torch.cuda.is_available()
-    if args.use_cuda:
-        print("Using GPU for acceleration")
-    else:
-        print("Using CPU for computation")
-
-    return args
-
-def deprocess_image(img):
-    """ see https://github.com/jacobgil/keras-grad-cam/blob/master/grad-cam.py#L65 """
-    img = img - np.mean(img)
-    img = img / (np.std(img) + 1e-5)
-    img = img * 0.1
-    img = img + 0.5
-    img = np.clip(img, 0, 1)
-    return np.uint8(img*255)
-
-if __name__ == '__main__':
-    """ python grad_cam.py <path_to_image>
-    1. Loads an image with opencv.
-    2. Preprocesses it for VGG19 and converts to a pytorch variable.
-    3. Makes a forward pass to find the category index with the highest score,
-    and computes intermediate activations.
-    Makes the visualization. """
-
-    args = get_args()
-
-    model = models.resnet50(pretrained=True)
-    grad_cam = GradCam(model=model, feature_module=model.layer4, \
-                       target_layer_names=["2"], use_cuda=args.use_cuda)
-
-    img = cv2.imread(args.image_path, 1)
-    img = np.float32(img) / 255
-    # Opencv loads as BGR:
-    img = img[:, :, ::-1]
-    input_img = preprocess_image(img)
-
-    # If None, returns the map for the highest scoring category.
-    # Otherwise, targets the requested category.
-    target_category = None
-    grayscale_cam = grad_cam(input_img, target_category)
-
-    grayscale_cam = cv2.resize(grayscale_cam, (img.shape[1], img.shape[0]))
-    cam = show_cam_on_image(img, grayscale_cam)
-
-    gb_model = GuidedBackpropReLUModel(model=model, use_cuda=args.use_cuda)
-    gb = gb_model(input_img, target_category=target_category)
-    gb = gb.transpose((1, 2, 0))
-
-    cam_mask = cv2.merge([grayscale_cam, grayscale_cam, grayscale_cam])
-    cam_gb = deprocess_image(cam_mask*gb)
-    gb = deprocess_image(gb)
-
-    cv2.imwrite("cam.jpg", cam)
-    cv2.imwrite('gb.jpg', gb)
-    cv2.imwrite('cam_gb.jpg', cam_gb)
