+++ setup.cfg
@@ -1,8 +1,8 @@
 [metadata]
-name = pytorch-grad-cam-jacobgil
+name = pytorch-grad-cam
 version = 1.0.0
-author = Jacob Gildenblat
-author_email = jacob.gildenblat@gmail.com
+author = Josep de Cid
+author_email = josep.de.cid@gmail.com
 description = Grad-CAM visualization tool for PyTorch models.
 long_description = file: README.md
 long_description_content_type = text/markdown
